+++ app/src/main/java/com/bangkit/capstone/beangreader/presentation/component/EmptyScreen.kt
@@ -1,33 +1,36 @@
 package com.bangkit.capstone.beangreader.presentation.component
 
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import com.airbnb.lottie.compose.LottieAnimation
-import com.airbnb.lottie.compose.LottieCompositionSpec
-import com.airbnb.lottie.compose.rememberLottieComposition
+import androidx.compose.ui.res.painterResource
+import com.bangkit.capstone.beangreader.R
 
 @Composable
 fun EmptyScreen(
     modifier: Modifier = Modifier
 ) {
-    val composition by rememberLottieComposition(LottieCompositionSpec.Asset("lottie/notfound.json"))
-
     Column(
         modifier = modifier
             .fillMaxSize(),
         verticalArrangement = Arrangement.Center,
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
-        LottieAnimation(
-            modifier = Modifier
-                .fillMaxSize(1f),
-            composition = composition,
-            iterations = Int.MAX_VALUE
+        Image(painter = painterResource(R.drawable.data_not_found),
+            contentDescription = null,
+            modifier = Modifier.fillMaxWidth()
+        )
+        Text(
+            text = "Data tidak Ditemukan",
+            style = MaterialTheme.typography.bodyLarge,
+            color = MaterialTheme.colorScheme.primary
         )
     }
 }
