+++ app/src/main/java/com/bangkit/capstone/beangreader/presentation/screen/favorite/FavoriteScreen.kt
@@ -5,12 +5,8 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -28,7 +24,6 @@ import com.bangkit.capstone.beangreader.presentation.screen.favorite.component.L
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun FavoriteScreen(
-    onBackClick: () -> Unit,
     navigateToDetail: (Int, String, Int) -> Unit,
     modifier: Modifier = Modifier,
     viewModel: FavoriteViewModel = hiltViewModel()
@@ -39,14 +34,6 @@ fun FavoriteScreen(
         topBar = {
             CenterAlignedTopAppBar(
                 title = { Text(text = stringResource(R.string.favorite)) },
-                navigationIcon = {
-                    IconButton(onClick = onBackClick) {
-                        Icon(
-                            imageVector = Icons.Default.ArrowBack,
-                            contentDescription = "back"
-                        )
-                    }
-                },
                 modifier = modifier
             )
         }
