--- app/src/main/java/com/bangkit/capstone/beangreader/presentation/screen/history/HistoryScreen.kt
@@ -1,17 +1,0 @@
-package com.bangkit.capstone.beangreader.presentation.screen.history
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import com.bangkit.capstone.beangreader.presentation.component.EmptyScreen
-
-@Composable
-fun HistoryScreen() {
-    HistoryContent()
-}
-
-@Composable
-fun HistoryContent(
-    modifier: Modifier = Modifier
-) {
-    EmptyScreen(modifier = modifier)
-}
