+++ app/src/main/java/com/bangkit/capstone/beangreader/presentation/screen/splash/SplashScreen.kt
@@ -5,6 +5,7 @@ import androidx.compose.animation.core.tween
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.size
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
@@ -15,8 +16,11 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.bangkit.capstone.beangreader.R
+import com.bangkit.capstone.beangreader.ui.theme.BeanGreaderTheme
 import kotlinx.coroutines.delay
 
 @Composable
@@ -45,9 +49,18 @@ fun SplashScreen(
         modifier = modifier.fillMaxSize()
     ) {
         Image(
-            painter = painterResource(R.drawable.beanicon),
+            painter = painterResource(R.drawable.splashbean),
             contentDescription = "logo",
-            alpha = alphaAnim.value
+            alpha = alphaAnim.value,
+            modifier = Modifier.size(196.dp)
         )
+    }
+}
+
+@Preview
+@Composable
+fun SplashPreview() {
+    BeanGreaderTheme {
+        SplashScreen(onTimeOut = {})
     }
 }
