--- app/src/main/res/drawable/beanicon.xml
@@ -1,43 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="240dp"
-    android:height="240dp"
-    android:viewportWidth="240"
-    android:viewportHeight="240">
-  <path
-      android:pathData="M120,195C161.42,195 195,161.42 195,120C195,78.58 161.42,45 120,45C78.58,45 45,78.58 45,120C45,161.42 78.58,195 120,195Z"
-      android:fillColor="#FCCE6B"/>
-  <path
-      android:pathData="M120.5,108.75C119.75,108.75 119.25,108.25 119.25,107.5V74.25C119.25,73.5 119.75,73 120.5,73H151C151.75,73 152.25,73.5 152.25,74.25C152.25,75 151.75,75.5 151,75.5H121.75V107.5C121.75,108 121,108.75 120.5,108.75Z"
-      android:fillColor="#54321F"/>
-  <path
-      android:pathData="M149.75,71H145.5L144.25,65.25H150.75L149.75,71Z"
-      android:fillColor="#54321F"/>
-  <path
-      android:pathData="M120.5,77.75C122.43,77.75 124,76.18 124,74.25C124,72.32 122.43,70.75 120.5,70.75C118.57,70.75 117,72.32 117,74.25C117,76.18 118.57,77.75 120.5,77.75Z"
-      android:fillColor="#54321F"/>
-  <path
-      android:pathData="M153,92V95H150.5C149,105.75 142,114.5 132.25,118.75C128.5,120.25 124.5,121.25 120.25,121.25C105,121.25 92.25,110 90,95.25H87.75V92.25H153V92Z"
-      android:fillColor="#54321F"/>
-  <path
-      android:pathData="M150.5,94.75H153V91.75H103.25L129.25,117.75H133.25C142.5,113.75 149,105.25 150.5,94.75Z"
-      android:strokeAlpha="0.25"
-      android:fillColor="#D8D8D7"
-      android:fillAlpha="0.25"/>
-  <path
-      android:pathData="M161.5,121V118H133.25H129.5H78.5V121H83.25V171.75H78.5V174.75H161.5V171.75H157.75V121H161.5Z"
-      android:fillColor="#54321F"/>
-  <path
-      android:pathData="M146,136H97.75V157.25H146V136Z"
-      android:fillColor="#ffffff"/>
-  <path
-      android:pathData="M106.75,121H157.75V171.75L106.75,121Z"
-      android:strokeAlpha="0.25"
-      android:fillColor="#D8D8D7"
-      android:fillAlpha="0.25"/>
-  <path
-      android:pathData="M161.5,171.75V172H157.75V171.75H161.5Z"
-      android:fillColor="#FCCE6B"/>
-  <path
-      android:pathData="M123.5,149.75C125.16,149.75 126.5,148.41 126.5,146.75C126.5,145.09 125.16,143.75 123.5,143.75C121.84,143.75 120.5,145.09 120.5,146.75C120.5,148.41 121.84,149.75 123.5,149.75Z"
-      android:fillColor="#49200C"/>
-</vector>
