+++ app/src/main/res/drawable/camera_add.xml
@@ -1,0 +1,15 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="800dp"
+    android:height="800dp"
+    android:viewportWidth="48"
+    android:viewportHeight="48">
+  <path
+      android:pathData="M20.515,7C18.972,7 17.55,7.837 16.8,9.186L15.514,11.5H11.75C8.574,11.5 6,14.074 6,17.25V34.25C6,37.426 8.574,40 11.75,40H22.996C22.664,39.204 22.409,38.367 22.24,37.5H11.75C9.955,37.5 8.5,36.045 8.5,34.25V17.25C8.5,15.455 9.955,14 11.75,14H16.986L18.985,10.4C19.294,9.845 19.88,9.5 20.515,9.5H27.485C28.12,9.5 28.706,9.845 29.015,10.4L31.014,14H36.25C38.045,14 39.5,15.455 39.5,17.25V22.8C40.382,23.125 41.22,23.544 42,24.044V17.25C42,14.074 39.426,11.5 36.25,11.5H32.486L31.2,9.186C30.45,7.837 29.028,7 27.485,7H20.515Z"
+      android:fillColor="#212121"/>
+  <path
+      android:pathData="M24,17C27.528,17 30.522,19.283 31.586,22.453C30.771,22.674 29.988,22.973 29.245,23.34C28.541,21.114 26.459,19.5 24,19.5C20.962,19.5 18.5,21.962 18.5,25C18.5,27.642 20.362,29.848 22.845,30.379C22.552,31.151 22.329,31.959 22.187,32.794C18.642,31.972 16,28.795 16,25C16,20.582 19.582,17 24,17Z"
+      android:fillColor="#212121"/>
+  <path
+      android:pathData="M46,35C46,41.075 41.075,46 35,46C28.925,46 24,41.075 24,35C24,28.925 28.925,24 35,24C41.075,24 46,28.925 46,35ZM36,29C36,28.448 35.552,28 35,28C34.448,28 34,28.448 34,29V34H29C28.448,34 28,34.448 28,35C28,35.552 28.448,36 29,36H34V41C34,41.552 34.448,42 35,42C35.552,42 36,41.552 36,41V36H41C41.552,36 42,35.552 42,35C42,34.448 41.552,34 41,34H36V29Z"
+      android:fillColor="#212121"/>
+</vector>
