+++ app/src/main/res/values-in/strings.xml
@@ -73,7 +73,7 @@
     <string name="delete_account">Hapus Akun</string>
     <string name="moisture">Kadar air</string>
     <string name="message">Pesan</string>
-    <string name="description_scan">Silahkan unggah foto biji kopi anda, untuk mengetahui kualitas biji kopi tersebut</string>
+    <string name="description_scan">Silahkan unggah foto biji kopi anda, untuk mengetahui kualitas biji kopi tersebut.</string>
     <string name="title_scan">Scan Biji Kopimu!</string>
     <string name="upload">Unggah</string>
     <string name="loading_desc">Sedang memproses gambar\nSilahkan tunggu…</string>
